+++ Source/MQTTnet/Protocol/MqttControlPacketType.cs
@@ -15,7 +15,7 @@ namespace MQTTnet.Protocol
         PubComp = 7,
         Subscribe = 8,
         SubAck = 9,
-        Unsubscibe = 10,
+        Unsubscribe = 10,
         UnsubAck = 11,
         PingReq = 12,
         PingResp = 13,