+++ Source/MQTTnet/Server/Internal/DispatchApplicationMessageResult.cs
@@ -2,15 +2,27 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
+using System.Collections.Generic;
+using MQTTnet.Packets;
+
 namespace MQTTnet.Server
 {
     public sealed class DispatchApplicationMessageResult
     {
-        public DispatchApplicationMessageResult(int matchingSubscribersCount)
+        public DispatchApplicationMessageResult(int reasonCode, bool closeConnection, string reasonString, List<MqttUserProperty> userProperties)
         {
-            MatchingSubscribersCount = matchingSubscribersCount;
+            ReasonCode = reasonCode;
+            CloseConnection = closeConnection;
+            ReasonString = reasonString;
+            UserProperties = userProperties;
         }
 
-        public int MatchingSubscribersCount { get; }
+        public bool CloseConnection { get; }
+
+        public int ReasonCode { get; }
+
+        public string ReasonString { get; }
+
+        public List<MqttUserProperty> UserProperties { get; }
     }
 }